Renomenado as classes dto e movendo para os pacotes request e response

+++ src/main/java/br/com/g1bet/service/UsuarioService.java
@@ -1,18 +1,16 @@
 package br.com.g1bet.service;
 
 import br.com.g1bet.exceptions.CampoExistenteException;
-import br.com.g1bet.exceptions.CampoNullException;
+import br.com.g1bet.exceptions.IdNaoExisteException;
 import br.com.g1bet.mapper.UsuarioMapper;
 import br.com.g1bet.model.Usuario;
-import br.com.g1bet.model.dto.UsuarioRequest;
-import br.com.g1bet.model.dto.UsuarioResponse;
+import br.com.g1bet.dto.request.UsuarioRequest;
+import br.com.g1bet.dto.response.UsuarioResponse;
 import br.com.g1bet.repository.UsuarioRepository;
 import lombok.RequiredArgsConstructor;
 import org.hibernate.ObjectNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @RequiredArgsConstructor
 @Service
@@ -25,28 +23,25 @@ public class UsuarioService {
         if (repository.existsByEmail(usuarioRequest.getEmail())) {
             throw new CampoExistenteException("Esse email já existe!");
         }
-        if (usuarioRequest.getSenha() == null
-                || usuarioRequest.getEmail() == null
-                || usuarioRequest.getDataDeNascimento() == null
-                || usuarioRequest.getNome() == null) {
-            throw new CampoNullException("Campo obrigatório não pode ser nulo");
-        }
 
         Usuario usuario = UsuarioMapper.INSTANCE.toUsuario(usuarioRequest);
         return UsuarioResponse.toUsuarioResponse(repository.save(usuario));
     }
 
     public void deletar(Long id) {
+        if (!repository.existsById(id)) {
+            throw new IdNaoExisteException("ID nao encontrado");
+        }
         repository.deleteById(id);
     }
 
-    public Usuario findById(Long id) {
-        Optional<Usuario> usuario = repository.findById(id);
-        return usuario.orElseThrow(() -> new ObjectNotFoundException("Id: " + id, "Usuario não encontrado"));
+    public Usuario buscar(Long id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new ObjectNotFoundException("Id: " + id, "Usuario não encontrado"));
     }
 
     public Usuario atualizar(Usuario usuario, Long id) {
-        Usuario atualizarUsuario = findById(id);
+        Usuario atualizarUsuario = buscar(id);
         atualizarUsuario.setNome(usuario.getNome());
         atualizarUsuario.setCpf(usuario.getCpf());
         atualizarUsuario.setDataDeNascimento(usuario.getDataDeNascimento());
